requirements.txt
annotated-types==0.6.0
bcrypt==4.0.1
EasyProcess==1.1
ecdsa==0.18.0
entrypoint2==1.1
MouseInfo==0.1.3
mss==9.0.1
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
pyasn1==0.5.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic_core==2.14.3
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Quartz==10.0
pyperclip==1.8.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.11
PyQt5-sip==12.13.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.30
python-dotenv==1.0.0
python-jose==3.3.0
pytweening==1.0.7
rsa==4.9
rubicon-objc==0.4.7
six==1.16.0
typing_extensions==4.8.0
vidstream==0.0.14
dotenv==0.19.1
python-jose==3.3.0
bcrypt==3.2.0
